requirements.txt
gdown==5.2.0
kaggle==1.6.17
keras==3.4.1
PyYAML==6.0.2
tensorflow==2.17.0
mlflow==2.16.1